+++ common/fingerprints.py
@@ -1,3 +1,4 @@
+import os
 
 _FINGERPRINTS = {
   "ACURA ILX 2016 ACURAWATCH PLUS": {
@@ -12,6 +13,11 @@ _FINGERPRINTS = {
   },
   "HONDA ACCORD 2016 TOURING": {
     1024L: 5, 929L: 8, 1027L: 5, 773L: 7, 1601L: 8, 777L: 8, 1036L: 8, 398L: 3, 1039L: 8, 401L: 8, 145L: 8, 1424L: 5, 660L: 8, 661L: 4, 918L: 7, 985L: 3, 923L: 2, 542L: 7, 927L: 8, 800L: 8, 545L: 4, 420L: 8, 422L: 8, 808L: 8, 426L: 8, 1029L: 8, 432L: 7, 57L: 3, 316L: 8, 829L: 5, 1600L: 5, 1089L: 8, 1057L: 5, 780L: 8, 1088L: 8, 464L: 8, 1108L: 8, 597L: 8, 342L: 6, 983L: 8, 344L: 8, 804L: 8, 476L: 4, 1296L: 3, 891L: 8, 1125L: 8, 487L: 4, 892L: 8, 490L: 8, 871L: 8, 1064L: 7, 882L: 2, 884L: 8, 506L: 8, 507L: 1, 380L: 8, 1365L: 5
+  },
+  "HONDA CR-V 2016 TOURING": {
+    57L: 3, 145L: 8, 316L: 8, 340L: 8, 342L: 6, 344L: 8, 380L: 8, 398L: 3, 399L: 6, 401L: 8, 420L: 8, 422L: 8, 426L: 8, 432L: 7, 464L: 8, 474L: 5, 476L: 4, 487L: 4, 490L: 8, 493L: 3, 507L: 1, 542L: 7, 545L: 4, 597L: 8, 660L: 8, 661L: 4, 773L: 7, 777L: 8, 800L: 8, 804L: 8, 808L: 8, 882L: 2, 884L: 7, 888L: 8, 891L: 8, 892L: 8, 923L: 2, 929L: 8, 983L: 8, 985L: 3, 1024L: 5, 1027L: 5, 1029L: 8, 1033L: 5, 1036L: 8, 1039L: 8, 1057L: 5, 1064L: 7, 1108L: 8, 1125L: 8, 1296L: 8, 1365L: 5, 1424L: 5, 1600L: 5, 1601L: 8,
+    # sent messages
+    0x194: 4, 0x1fa: 8, 0x30c: 8, 0x33d: 5,
   }
 }
 
@@ -46,30 +52,13 @@ def fingerprint(logcan):
   import selfdrive.messaging as messaging
   from cereal import car
   from common.realtime import sec_since_boot
-  import os
+
   if os.getenv("SIMULATOR") is not None or logcan is None:
-    # send message
-    ret = car.CarParams.new_message()
-
-    ret.carName = "simulator"
-    ret.radarName = "nidec"
-    ret.carFingerprint = "THE LOW QUALITY SIMULATOR"
-
-    ret.enableSteer = True
-    ret.enableBrake = True
-    ret.enableGas = True
-    ret.enableCruise = False
-
-    ret.wheelBase = 2.67
-    ret.steerRatio = 15.3
-    ret.slipFactor = 0.0014
-
-    ret.steerKp, ret.steerKi = 12.0, 1.0
-    return ret
+    return ("simulator", None)
+  elif os.getenv("SIMULATOR2") is not None:
+    return ("simulator2", None)
 
   print "waiting for fingerprint..."
-  brake_only = True
-
   candidate_cars = all_known_cars()
   finger = {}
   st = None
@@ -78,9 +67,6 @@ def fingerprint(logcan):
       if st is None:
         st = sec_since_boot()
       for can in a.can:
-        # pedal
-        if can.address == 0x201 and can.src == 0:
-          brake_only = False
         if can.src == 0:
           finger[can.address] = len(can.dat)
         candidate_cars = eliminate_incompatible_cars(can, candidate_cars)
@@ -93,36 +79,4 @@ def fingerprint(logcan):
       raise Exception("car doesn't match any fingerprints")
 
   print "fingerprinted", candidate_cars[0]
-
-  # send message
-  ret = car.CarParams.new_message()
-
-  ret.carName = "honda"
-  ret.radarName = "nidec"
-  ret.carFingerprint = candidate_cars[0]
-
-  ret.enableSteer = True
-  ret.enableBrake = True
-  ret.enableGas = not brake_only
-  ret.enableCruise = brake_only
-  #ret.enableCruise = False
-
-  ret.wheelBase = 2.67
-  ret.steerRatio = 15.3
-  ret.slipFactor = 0.0014
-
-  if candidate_cars[0] == "HONDA CIVIC 2016 TOURING":
-    ret.steerKp, ret.steerKi = 12.0, 1.0
-  elif candidate_cars[0] == "ACURA ILX 2016 ACURAWATCH PLUS":
-    if not brake_only:
-      # assuming if we have an interceptor we also have a torque mod
-      ret.steerKp, ret.steerKi = 6.0, 0.5
-    else:
-      ret.steerKp, ret.steerKi = 12.0, 1.0
-  elif candidate_cars[0] == "HONDA ACCORD 2016 TOURING":
-    ret.steerKp, ret.steerKi = 12.0, 1.0
-  else:
-    raise ValueError("unsupported car %s" % candidate_cars[0])
-
-  return ret
-
+  return (candidate_cars[0], finger)