+++ common/logging_extra.py
@@ -8,6 +8,18 @@ from threading import local
 from collections import OrderedDict
 from contextlib import contextmanager
 
+def json_handler(obj):
+  # if isinstance(obj, (datetime.date, datetime.time)):
+  #   return obj.isoformat()
+  return repr(obj)
+
+def json_robust_dumps(obj):
+  return json.dumps(obj, default=json_handler)
+
+class NiceOrderedDict(OrderedDict):
+  def __str__(self):
+    return '{'+', '.join("%r: %r" % p for p in self.iteritems())+'}'
+
 class SwagFormatter(logging.Formatter):
   def __init__(self, swaglogger):
     logging.Formatter.__init__(self, None, '%a %b %d %H:%M:%S %Z %Y')
@@ -15,13 +27,8 @@ class SwagFormatter(logging.Formatter):
     self.swaglogger = swaglogger
     self.host = socket.gethostname()
 
-  def json_handler(self, obj):
-    # if isinstance(obj, (datetime.date, datetime.time)):
-    #   return obj.isoformat()
-    return repr(obj)
-
-  def format(self, record):
-    record_dict = OrderedDict()
+  def format_dict(self, record):
+    record_dict = NiceOrderedDict()
 
     if isinstance(record.msg, dict):
       record_dict['msg'] = record.msg
@@ -50,9 +57,10 @@ class SwagFormatter(logging.Formatter):
     record_dict['threadName'] = record.threadName
     record_dict['created'] = record.created
 
-    # asctime = self.formatTime(record, self.datefmt)
+    return record_dict
 
-    return json.dumps(record_dict, default=self.json_handler)
+  def format(self, record):
+    return json_robust_dumps(self.format_dict(record))
 
 _tmpfunc = lambda: 0
 _srcfile = os.path.normcase(_tmpfunc.__code__.co_filename)
@@ -115,7 +123,7 @@ class SwagLogger(logging.Logger):
     self.global_ctx.update(kwargs)
 
   def event(self, event_name, *args, **kwargs):
-    evt = OrderedDict()
+    evt = NiceOrderedDict()
     evt['event'] = event_name
     if args:
       evt['args'] = args