+++ selfdrive/loggerd/uploader.py
@@ -117,7 +117,12 @@ class Uploader(object):
       if name in ["rlog", "rlog.bz2"]:
         return (key, fn, 0)
 
-    # then upload camera files no not on wifi
+    # then upload compressed camera file
+    for name, key, fn in self.gen_upload_files():
+      if name in ["fcamera.hevc"]:
+        return (key, fn, 1)
+
+    # then upload other files
     for name, key, fn in self.gen_upload_files():
       if not name.endswith('.lock') and not name.endswith(".tmp"):
         return (key, fn, 1)
@@ -131,10 +136,10 @@ class Uploader(object):
       url_resp_json = json.loads(url_resp.text)
       url = url_resp_json['url']
       headers = url_resp_json['headers']
-      cloudlog.info({"upload_url v1.1", url, str(headers)})
+      cloudlog.info("upload_url v1.1 %s %s", url, str(headers))
 
       if fake_upload:
-        print "*** WARNING, THIS IS A FAKE UPLOAD TO %s ***" % url
+        cloudlog.info("*** WARNING, THIS IS A FAKE UPLOAD TO %s ***" % url)
         class FakeResponse(object):
           def __init__(self):
             self.status_code = 200